+++ core-bundle/src/Resources/contao/elements/ContentElement.php
@@ -68,6 +68,7 @@ namespace Contao;
  * @property string  $customTpl
  * @property string  $playerSRC
  * @property string  $youtube
+ * @property string  $vimeo
  * @property string  $posterSRC
  * @property string  $playerSize
  * @property boolean $autoplay